Fix garbled error message in debug log
CCOBOL source code pre-compiled with the ocesql command of Open Cobol ESQL
for the COBOL to C translator will produce an error at runtime when compiled
using the cobj command of opensource COBOL 4J which is a COBOL to Java
transrator.

At this time, if you set "debug" in the OCDB_LOGLEVEL environment variable,
the log will be output to the /tmp/ocesql.log file, but the error message is
garbled.

This commit fixes garbled error message in debug log.

Signed-off-by: Masayuki Moriyama <masayuki.moriyama@miraclelinux.com>